[AOS] refactor: 패키지 구조화
