+++ dodamduck_aos/app/src/main/java/org/chosun/dodamduck/presentation/chat/ChatViewModel.kt
@@ -1,4 +1,4 @@
-package org.chosun.dodamduck.model.viewmodel
+package org.chosun.dodamduck.presentation.chat
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
