+++ dodamduck_aos/app/src/main/java/org/chosun/dodamduck/presentation/post/PostScreen.kt
@@ -1,6 +1,7 @@
-package org.chosun.dodamduck.ui
+package org.chosun.dodamduck.presentation.post
 
 import androidx.compose.foundation.BorderStroke
+import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
@@ -11,11 +12,12 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Add
-import androidx.compose.material.icons.filled.KeyboardArrowDown
 import androidx.compose.material.icons.filled.Search
 import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.Divider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.OutlinedButton
 import androidx.compose.material3.Text
@@ -23,9 +25,14 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
@@ -33,22 +40,30 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.rememberNavController
 import org.chosun.dodamduck.R
-import org.chosun.dodamduck.model.viewmodel.TradeViewModel
+import org.chosun.dodamduck.model.dto.CategoryDTO
 import org.chosun.dodamduck.ui.component.DodamDuckTextH2
-import org.chosun.dodamduck.ui.component.lazy_components.ExchangeItemList
+import org.chosun.dodamduck.ui.component.lazy_components.PostItem
+import org.chosun.dodamduck.ui.component.lazy_components.TagLazyRow
 import org.chosun.dodamduck.ui.navigation.BottomNavItem
 import org.chosun.dodamduck.ui.theme.Brown
 import org.chosun.dodamduck.ui.theme.DodamDuckTheme
 
 @Composable
-fun TradeScreen(
+fun PostScreen(
     navController: NavHostController,
-    tradeViewModel: TradeViewModel = hiltViewModel()
+    postViewModel: PostViewModel = hiltViewModel()
 ) {
-    val tradeLists by tradeViewModel.postLists.collectAsState(initial = null)
+    val postLists by postViewModel.postLists.collectAsState(initial = null)
+    val categories by postViewModel.categories.collectAsState(initial = null)
+    var selectedTag by remember { mutableStateOf(CategoryDTO("0", "")) }
 
     LaunchedEffect(Unit) {
-        tradeViewModel.getTradeLists()
+        postViewModel.getPostLists()
+        postViewModel.getCategories()
+    }
+
+    LaunchedEffect(key1 = selectedTag) {
+        postViewModel.getPostLists(selectedTag.id)
     }
 
     Box(
@@ -57,12 +72,47 @@ fun TradeScreen(
             .background(Color.White)
     ) {
         Column {
-            TradeHeader(navController)
-            ExchangeItemList(
-                modifier = Modifier.padding(top = 24.dp),
-                items = tradeLists ?: listOf(),
-                navController = navController
+            PostHeader(navController)
+            Divider(modifier = Modifier.padding(top = 12.dp))
+
+            TagLazyRow(
+                Modifier.padding(start = 8.dp, top = 10.dp),
+                categories = categories ?: listOf(),
+                selectedTag = selectedTag,
+                onTagSelected = { selectedTag = it }
             )
+            Divider(modifier = Modifier.padding(top = 10.dp))
+
+            val posts = postLists ?: listOf()
+            LazyColumn {
+                items(posts.size) {index ->
+                    PostItem(modifier = Modifier
+                        .padding(start = 8.dp, end = 8.dp, top = 8.dp)
+                        .clickable {
+                            postViewModel.uploadViewCount(posts[index].shareID)
+                            navController.navigate(
+                                "${BottomNavItem.PostDetail.screenRoute}/${posts[index].shareID}/post"
+                            ) {
+                                popUpTo(navController.graph.startDestinationId) {
+                                    saveState = true
+                                }
+                                launchSingleTop = true
+                            }
+                        },
+                        item = posts[index]
+                    )
+                    Divider()
+                }
+            }
+
+            if(posts.isEmpty()) {
+                Image(
+                    modifier = Modifier.fillMaxSize(),
+                    painter = painterResource(id = R.drawable.img_no_post),
+                    contentDescription = "No Post Image",
+                    contentScale = ContentScale.FillWidth
+                )
+            }
         }
 
         OutlinedButton(
@@ -70,7 +120,7 @@ fun TradeScreen(
                 .align(Alignment.BottomEnd)
                 .height(60.dp)
                 .padding(end = 8.dp, bottom = 8.dp),
-            onClick = { navController.navigate(BottomNavItem.TradeWrite.screenRoute) },
+            onClick = { navController.navigate(BottomNavItem.PostWrite.screenRoute) },
             colors = ButtonDefaults.outlinedButtonColors(contentColor = Brown),
             border = BorderStroke(width = 1.dp, color = Brown)
         ) {
@@ -82,14 +132,15 @@ fun TradeScreen(
 }
 
 @Composable
-fun TradeHeader(navController: NavHostController) {
+fun PostHeader(
+    navController: NavHostController
+) {
     Row(
         modifier = Modifier
             .fillMaxWidth()
             .padding(start = 10.dp, end = 10.dp, top = 12.dp)
     ) {
         DodamDuckTextH2(text = "빛가람동", color = Brown)
-        Icon(imageVector = Icons.Default.KeyboardArrowDown, contentDescription = "Arrow Icon")
         Spacer(modifier = Modifier.weight(1f))
         Icon(
             modifier = Modifier.clickable { navController.navigate(BottomNavItem.Search.screenRoute) },
@@ -101,8 +152,8 @@ fun TradeHeader(navController: NavHostController) {
 
 @Preview
 @Composable
-fun TradePreview() {
+fun PostPreview() {
     DodamDuckTheme {
-        TradeScreen(rememberNavController())
+        PostScreen(rememberNavController())
     }
 }
