+++ dodamduck_aos/app/src/main/java/org/chosun/dodamduck/ui/navigation/DodamDuckNavigation.kt
@@ -22,18 +22,18 @@ import androidx.navigation.compose.composable
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.navArgument
 import org.chosun.dodamduck.R
-import org.chosun.dodamduck.ui.ChatListScreen
-import org.chosun.dodamduck.ui.ChatScreen
-import org.chosun.dodamduck.ui.PostScreen
-import org.chosun.dodamduck.ui.TradeScreen
-import org.chosun.dodamduck.ui.LibraryScreen
-import org.chosun.dodamduck.ui.LoginScreen
-import org.chosun.dodamduck.ui.OnboardingScreen
-import org.chosun.dodamduck.ui.PostDetailScreen
-import org.chosun.dodamduck.ui.PostWriteScreen
-import org.chosun.dodamduck.ui.RegisterScreen
-import org.chosun.dodamduck.ui.SearchScreen
-import org.chosun.dodamduck.ui.TradeWriteScreen
+import org.chosun.dodamduck.presentation.chat.ChatListScreen
+import org.chosun.dodamduck.presentation.chat.ChatScreen
+import org.chosun.dodamduck.presentation.post.PostScreen
+import org.chosun.dodamduck.presentation.trade.TradeScreen
+import org.chosun.dodamduck.presentation.library.LibraryScreen
+import org.chosun.dodamduck.presentation.auth.LoginScreen
+import org.chosun.dodamduck.presentation.auth.OnboardingScreen
+import org.chosun.dodamduck.presentation.post.PostDetailScreen
+import org.chosun.dodamduck.presentation.post.PostWriteScreen
+import org.chosun.dodamduck.presentation.auth.RegisterScreen
+import org.chosun.dodamduck.presentation.search.SearchScreen
+import org.chosun.dodamduck.presentation.trade.TradeWriteScreen
 import org.chosun.dodamduck.ui.UserScreen
 import org.chosun.dodamduck.ui.theme.Brown
